+++ siade/app/controllers/api_entreprise/v3_and_more/mi/siaf/associations_open_data_controller.rb
@@ -1,0 +1,7 @@
+class APIEntreprise::V3AndMore::MI::SIAF::AssociationsOpenDataController < APIEntreprise::V3AndMore::MI::SIAF::AssociationsController
+  private
+
+  def serializer_module
+    ::APIEntreprise::MI::SIAF::AssociationsOpenDataSerializer
+  end
+end