+++ siade/config/routes/api_entreprise.rb
@@ -98,6 +98,8 @@ scope path: 'v:api_version', constraints: APIEntrepriseDomainConstraint.new(v3_a
   get 'djepva/api-association/associations/open_data/:siren_or_rna', to: 'api_entreprise/v3_and_more/mi/unites_legales_open_data#show'
 
   get 'ministere_interieur/siaf/fondations/:siren_or_siret_or_rnf', to: 'api_entreprise/v3_and_more/mi/siaf/fondations#show'
+  get 'ministere_interieur/siaf/associations/open_data/:siren_or_siret_or_rna', to: 'api_entreprise/v3_and_more/mi/siaf/associations_open_data#show'
+  get 'ministere_interieur/siaf/associations/:siren_or_siret_or_rna', to: 'api_entreprise/v3_and_more/mi/siaf/associations#show'
 
   get 'data_subvention/associations/:siren_or_siret_or_rna/subventions', to: 'api_entreprise/v3_and_more/data_subvention/subventions#show'